feat: add decision for modules task
